Refactor Secured component: moved from components to views directory, updated access token display to show authentication status, and removed unnecessary imports. This enhances the organization of the project and improves user feedback on authentication.
